fix unsigned

+++ csrc/batch_prefill_sm90.cu
@@ -40,7 +40,7 @@ using namespace flashinfer;
 at::Tensor BatchPrefillWithKVCacheSM90Plan(
     at::Tensor float_workspace_buffer, at::Tensor int_workspace_buffer,
     at::Tensor page_locked_int_workspace_buffer, at::Tensor qo_indptr, at::Tensor kv_indptr,
-    at::Tensor kv_len_arr, unsigned total_num_rows, int64_t batch_size,
+    at::Tensor kv_len_arr, int64_t total_num_rows, int64_t batch_size,
     int64_t num_qo_heads, int64_t num_kv_heads, int64_t page_size,
     bool enable_cuda_graph, int64_t head_dim, bool causal, int64_t cuda_stream) {
   size_t float_workspace_size_in_bytes =
